requirements.txt
numpy==1.24.3
Pillow==9.4.0
Pillow==10.0.0
Pillow==10.1.0
